+++ src/Model/Question.php
@@ -6,24 +6,8 @@ namespace ThreeBRS\SyliusShipmentExportPlugin\Model;
 
 class Question implements QuestionInterface
 {
-    /** @var string */
-    private $code;
-
-    /** @var string */
-    private $label;
-
-    /** @var string|null */
-    private $defaultValue;
-
-    /** @var string|null */
-    private $regex;
-
-    public function __construct(string $code, string $label, ?string $defaultValue, ?string $regex)
+    public function __construct(private string $code, private string $label, private ?string $defaultValue, private ?string $regex)
     {
-        $this->code = $code;
-        $this->label = $label;
-        $this->defaultValue = $defaultValue;
-        $this->regex = $regex;
     }
 
     public function getCode(): string